EdgeDet.py and EdgeDetRT.py
Uploaded new files EdgeDet.py and EdgeDetRT.py. 

EdgeDet.py takes in a image an calculates/detects the major edges in the image and saves it as a image in the currently working directory. 

EdgeDetRT.py utilizes the constant input of the camera feed to calculate/detect major edges within each frame of the video capture. Displaying a real-time detection of major edges within the cameras field of view.